PremiumRoute/requirements.txt
# Core Django & Database
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-allauth==0.54.0
django-environ==0.10.0
psycopg2-binary==2.9.7
Pillow==10.0.1

# API & Security
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Async & Background Tasks
celery==5.3.1
redis==4.6.0
django-celery-results==2.5.1
django-celery-beat==2.5.0

# Payments
stripe==6.4.0

# Reporting & Documents
reportlab==4.0.4
django-import-export==3.3.0

# Filtering & Search
django-filter==23.3

# Static Files
whitenoise==6.5.0

# Production Server
gunicorn==21.2.0

# Email & SMS
twilio==8.8.0
django-anymail==10.2

# File Handling
django-storages==1.14.2
boto3==1.34.0

# Utilities
python-dateutil==2.8.2
pandas==2.1.4
qrcode==7.4.2
faker==20.1.0

# Country & Phone
django-countries==7.5.1
django-phonenumber-field==7.2.0
phonenumbers==8.13.28

# Security
django-axes==6.0.0
django-defender==1.1.1

# Monitoring
django-health-check==3.19.0

# Development Only (separate requirements-dev.txt)
django-debug-toolbar==4.2.0
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2